+++ eslint.config.js
@@ -3,6 +3,8 @@
 const path = require('path');
 const { includeIgnoreFile } = require('@eslint/compat');
 const prettierConfig = require('eslint-config-prettier');
+const tseslint = require('@typescript-eslint/eslint-plugin');
+const tsParser = require('@typescript-eslint/parser');
 
 const prettierIgnorePath = path.resolve(__dirname, '.prettierignore');
 
@@ -17,6 +19,24 @@ module.exports = [
         },
         rules: {
             'no-unused-vars': 'warn',
+            'no-console': 'off',
+        },
+    },
+    {
+        files: ['**/*.ts'],
+        plugins: {
+            '@typescript-eslint': tseslint,
+        },
+        languageOptions: {
+            parser: tsParser,
+            parserOptions: {
+                ecmaVersion: 'latest',
+                sourceType: 'module',
+            },
+        },
+        rules: {
+            ...tseslint.configs.recommended.rules,
+            '@typescript-eslint/no-unused-vars': 'warn',
             'no-console': 'off',
         },
     },